+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Upcoming changes compared to [NetCoreDbg](https://github.com/Samsung/netcoredbg)
 - Added methods parameters output in stacktrace.
 - Added active CLR internal frames output in stacktrace.
 - Added proper Just My Code enabled stacktrace.
+- Added `hitBreakpointIds` support in Stopped Event.
 
 #### Changed
 - Replaced VSCode to DAP (variables, class names, tests, etc).